spinner: Introduce PhocSpinner
The initial groundwork for an animated spinner. There's no rendering
yet, that comes later.

To begin, the animation bits are put in place. A spinner is assumed to
be a static sprite that is rotated all 360 degrees around its center
point, so rotation is the property we're animating here.

The animation bits are wired up, but nothing is playing it yet. That
also comes later.

Signed-off-by: Sam Day <me@samcday.com>