output-shield: Render a PhocColorRect
Instead of managing an alpha prop and drawing a rect directly, use
PhocColorRect. This removes some duplicated code and a TODO.

Today, PhocBling only has machinery in place to trivialize its
attachment to the render tree via a PhocView. That's not appropriate for
our use-case here, so we just map/unmap/render the bling via its
interface directly.

It's expected that eventually broader support for PhocBling rendering
will exist. This will simplify PhocOutputShield further, as it will then
no longer need to attach to the render pass.

Signed-off-by: Sam Day <me@samcday.com>